+++ sperl_field.c
@@ -11,10 +11,10 @@ SPerl_FIELD* SPerl_FIELD_new(SPerl* sperl) {
   return SPerl_ALLOCATOR_PARSER_alloc_memory_pool(sperl, sperl->parser, sizeof(SPerl_FIELD));
 }
 
-int32_t SPerl_FIELD_get_byte_size(SPerl* sperl, SPerl_FIELD* field) {
+int64_t SPerl_FIELD_get_byte_size(SPerl* sperl, SPerl_FIELD* field) {
   SPerl_RESOLVED_TYPE* field_resolved_type = field->op_type->uv.type->resolved_type;
   
-  int32_t byte_size;
+  int64_t byte_size;
   if (field_resolved_type->id == SPerl_RESOLVED_TYPE_C_ID_BYTE) {
     byte_size = 1;
   }
